RaspberryPi RFID Service/requirements.txt
RPi.GPIO
spidev
pi-rc522